demo_e2e/requirements.txt
pyrogram==2.0.106
pytest==8.4.2
python-dotenv==1.0.1